- Fixed hard coded path reference to if image in fluid about_panel.fl - Fixed minor warnings due to incorrect casts - Added TODO file to ide/visualc project
git-svn-id: http://seriss.com/public/fltk/fltk/branches/fctrunk@5752 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ fltk/AnsiWidget.h
@@ -32,7 +32,7 @@ using namespace fltk;
 
 class AnsiWidget : public Widget {
     public:
-    AnsiWidget(int x, int y, int w, int h, int defsize);
+    AnsiWidget(int x, int y, int w, int h, float defsize);
     virtual ~AnsiWidget();
 
     // inherited methods
@@ -59,7 +59,7 @@ class AnsiWidget : public Widget {
     int textHeight(void);
     int getWidth()  {return w();}
     int getHeight() {return h();}
-    void setFontSize(int i) {labelsize(i);}
+    void setFontSize(float i) {labelsize(i);}
     int getFontSize() {return (int)labelsize();}
     void beep() const;
     static Color ansiToFltk(long color);
